+++ shoop/admin/modules/orders/views/create.py
@@ -12,7 +12,6 @@ import json
 
 from babel.numbers import format_decimal
 from django.contrib import messages
-from django.contrib.auth.models import AnonymousUser
 from django.core import serializers
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.core.urlresolvers import reverse
@@ -27,8 +26,8 @@ from django_countries import countries
 
 from shoop.admin.modules.orders.json_order_creator import JsonOrderCreator
 from shoop.core.models import (
-    CompanyContact, Contact, MethodStatus, Order, PaymentMethod, Product,
-    ShippingMethod, Shop, ShopStatus
+    AnonymousContact, CompanyContact, Contact, MethodStatus, Order,
+    PaymentMethod, Product, ShippingMethod, Shop, ShopStatus
 )
 from shoop.core.pricing import get_pricing_module
 from shoop.utils.i18n import (
@@ -97,9 +96,7 @@ def encode_line(line):
 def get_price_info(shop, customer, product, quantity):
     ctx_request = RequestFactory().get("/")
     ctx_request.shop = shop
-    if customer:
-        ctx_request.customer = customer
-    ctx_request.user = AnonymousUser()
+    ctx_request.customer = (customer or AnonymousContact())
     context = get_pricing_module().get_context_from_request(ctx_request)
     return product.get_price_info(context, quantity=quantity)