+++ shoop/core/pricing/_context.py
@@ -7,11 +7,9 @@
 from __future__ import unicode_literals
 
 import abc
-import hashlib
 
 import six
 from django.http import HttpRequest
-from django.utils.encoding import force_bytes
 from django.utils.timezone import now
 
 
@@ -41,24 +39,14 @@ class PricingContext(PricingContextable):
     """
     Context for pricing.
     """
-    REQUIRED_VALUES = ()
+    def __init__(self, shop, customer, time=None):
+        """
+        Initialize pricing context for shop and customer.
 
-    def __init__(self, **kwargs):
-        kwargs.setdefault("time", now())
-        for name, value in kwargs.items():
-            setattr(self, name, value)
-        for name in self.REQUIRED_VALUES:
-            if not hasattr(self, name):
-                raise ValueError("%s is a required value for %s but is not set." % (name, self))
-
-    def get_cache_key_parts(self):
-        return [getattr(self, key) for key in self.REQUIRED_VALUES]
-
-    def get_cache_key(self):
-        parts_text = "\n".join(force_bytes(part) for part in self.get_cache_key_parts())
-        return "%s_%s" % (
-            self.__class__.__name__,
-            hashlib.sha1(parts_text).hexdigest()
-        )
-
-    cache_key = property(get_cache_key)
+        :type shop: shoop.core.models.Shop
+        :type customer: shoop.core.models.Contact
+        :type time: datetime.datetime|None
+        """
+        self.shop = shop
+        self.customer = customer
+        self.time = (time if time is not None else now())