+++ shoop/simple_pricing/module.py
@@ -10,35 +10,21 @@ from django.db.models import Q
 from django.utils.translation import ugettext_lazy as _
 
 from shoop.core.models import ShopProduct
-from shoop.core.pricing import PriceInfo, PricingContext, PricingModule
+from shoop.core.pricing import PriceInfo, PricingModule
 
 from .models import SimpleProductPrice
 
 
-class SimplePricingContext(PricingContext):
-    REQUIRED_VALUES = ("customer_group_ids", "shop")
-    customer_group_ids = ()
-    shop = None
+def _get_customer_group_ids(customer):
+    # TODO: add caching of some sort
+    if customer.is_anonymous:
+        return set()
+    return set(customer.groups.all().values_list("id", flat=True))
 
 
 class SimplePricingModule(PricingModule):
     identifier = "simple_pricing"
     name = _("Simple Pricing")
-
-    pricing_context_class = SimplePricingContext
-
-    def get_context_from_request(self, request):
-        customer = getattr(request, "customer", None)
-
-        if not customer or customer.is_anonymous:
-            customer_group_ids = []
-        else:
-            customer_group_ids = sorted(customer.groups.all().values_list("id", flat=True))
-
-        return self.pricing_context_class(
-            shop=request.shop,
-            customer_group_ids=customer_group_ids
-        )
 
     def get_price_info(self, context, product, quantity=1):
         shop = context.shop
@@ -52,10 +38,12 @@ class SimplePricingModule(PricingModule):
 
         default_price = (shop_product.default_price_value or 0)
 
-        if context.customer_group_ids:
+        customer_group_ids = _get_customer_group_ids(context.customer)
+
+        if customer_group_ids:
             filter = Q(
                 price_value__gt=0, product=product_id, shop=shop,
-                group__in=context.customer_group_ids)
+                group__in=customer_group_ids)
             result = (
                 SimpleProductPrice.objects.filter(filter)
                 .order_by("price_value")[:1]